add new parseAndSanitize method, move sanitize to same file
